move components back to tapis-ui and decompose layouts
