--- src/tapis-app/pages/Apps/pages/JobLauncher/components/JobLauncher/index.ts
@@ -1,3 +1,0 @@
-import { default as JobLauncher } from './JobLauncher';
-
-export default JobLauncher;
