--- src/tapis-app/pages/Apps/pages/JobLauncher/components/index.ts
@@ -1,1 +1,0 @@
-export { default as JobLauncher } from './JobLauncher';
