+++ src/tapis-app/pages/Jobs/Layout/Layout.tsx
@@ -15,10 +15,6 @@ const Layout: React.FC = () => {
     </LayoutHeader>
   );
 
-  const subHeader = (
-    <LayoutHeader type={'sub-header'}>Job Details</LayoutHeader>
-  );
-
   const sidebar = (
     <LayoutNavWrapper>
       <JobsNav />
@@ -26,22 +22,12 @@ const Layout: React.FC = () => {
   );
 
   const body = (
-    <div style={{ flex: 1 }}>
+    <LayoutBody>
       <Router />
-    </div>
+    </LayoutBody>
   );
 
-  return (
-    <PageLayout
-      top={header}
-      left={sidebar}
-      right={
-        <LayoutBody>
-          <PageLayout top={subHeader} right={body} />
-        </LayoutBody>
-      }
-    />
-  );
+  return <PageLayout top={header} left={sidebar} right={body} />;
 };
 
 export default Layout;
