+++ src/tapis-app/pages/Jobs/pages/JobDetail/Layout/Layout.tsx
@@ -1,24 +1,20 @@
 import React from 'react';
-import { useDetails } from 'tapis-hooks/jobs';
-import { Jobs } from '@tapis/tapis-typescript';
-import { DescriptionList } from 'tapis-ui/_common';
-import { QueryWrapper } from 'tapis-ui/_wrappers';
+import { JobDetail } from 'tapis-ui/components/jobs';
+import { PageLayout, LayoutHeader } from 'tapis-ui/_common';
 
 interface JobDetailProps {
   jobUuid: string;
 }
 
 const Layout: React.FC<JobDetailProps> = ({ jobUuid }) => {
-  const { data, isLoading, error } = useDetails(jobUuid);
-  const job: Jobs.Job | undefined = data?.result;
+  const header = <LayoutHeader type={'sub-header'}>Job Details</LayoutHeader>;
 
-  return (
-    <QueryWrapper isLoading={isLoading} error={error}>
-      <h3>{job?.name}</h3>
-      <h5>{job?.uuid}</h5>
-      {job && <DescriptionList data={job} />}
-    </QueryWrapper>
+  const body = (
+    <div style={{ flex: 1 }}>
+      <JobDetail jobUuid={jobUuid}></JobDetail>
+    </div>
   );
+  return <PageLayout top={header} right={body}></PageLayout>;
 };
 
 export default Layout;
