+++ src/tapis-app/pages/Systems/Layout/Layout.tsx
@@ -15,8 +15,6 @@ const Layout: React.FC = () => {
     </LayoutHeader>
   );
 
-  const subHeader = <LayoutHeader type={'sub-header'}>Files</LayoutHeader>;
-
   const sidebar = (
     <LayoutNavWrapper>
       <SystemsNav />
@@ -24,22 +22,12 @@ const Layout: React.FC = () => {
   );
 
   const body = (
-    <div style={{ flex: 1 }}>
+    <LayoutBody>
       <Router />
-    </div>
+    </LayoutBody>
   );
 
-  return (
-    <PageLayout
-      top={header}
-      left={sidebar}
-      right={
-        <LayoutBody>
-          <PageLayout top={subHeader} right={body} />
-        </LayoutBody>
-      }
-    />
-  );
+  return <PageLayout top={header} left={sidebar} right={body} />;
 };
 
 export default Layout;
