+++ src/tapis-ui/components/apps/JobLauncher/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './JobLauncher';
