+++ src/tapis-ui/components/jobs/JobDetail/index.ts
@@ -1,0 +1,1 @@
+export { default } from './JobDetail';
