+++ src/tapis-ui/components/jobs/index.ts
@@ -1,0 +1,1 @@
+export { default as JobDetail } from './JobDetail';
